app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
click==8.1.7
fastapi==0.115.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
